requirements.txt
Flask==3.1.1
keras==3.10.0
numpy
Pillow==11.2.1
tensorflow==2.19.0
